move function back
